==> legacy/05_vicinity_merges.ipynb <==
from collections import namedtuple
import numpy as np

import pandas as pd

Best = namedtuple("Best", ["src_1", "src_2", "n_1", "n_2", "int_12", "mmjd_1", "mmjd_2", "mmjd_diff"])
fields = list(Best._fields)

def select_nbg(df):
    cleaned = []
    while len(df):
        selected = df.sort_values(["mmjd_diff", "int_12"]).iloc[0]
        rm_key = map(int, [selected.src_1, selected.src_2])
        df = df[~(df.src_1.isin(rm_key)) & ~(df.src_2.isin(rm_key))]
        cleaned.append(selected.to_dict())
    try:
        return pd.DataFrame(cleaned)[fields]
    except:
        pass

def select_no_nbg(df):
    cleaned = []
    while len(df):
        selected = df.sort_values(["mmjd_diff", "int_12"], ascending=False).iloc[0]
        rm_key = map(int, [selected.src_1, selected.src_2])
        df = df[~(df.src_1.isin(rm_key)) & ~(df.src_2.isin(rm_key))]
        cleaned.append(selected.to_dict())
    try:
        return pd.DataFrame(cleaned)[fields]
    except:
        pass

nbgs, no_nbgs = [], []
for idx in ["", "2", "3"]:
    no_df = pd.read_pickle("data/no_nbgs{}.pkl".format(idx))
    df = pd.read_pickle("data/nbgs{}.pkl".format(idx))
    nbgs.append(df)
    no_nbgs.append(no_df)

print map(len, nbgs)
print map(len, no_nbgs)

nbgs = pd.concat(nbgs)
no_nbgs = pd.concat(no_nbgs)

%time nbgs = select_nbg(nbgs)

%time no_nbgs = select_no_nbg(no_nbgs)

print len(nbgs)
print len(no_nbgs)

np.unique(nbgs[["src_1", "src_2"]].values.flatten()).shape == nbgs[["src_1", "src_2"]].values.flatten().shape

np.unique(no_nbgs[["src_1", "src_2"]].values.flatten()).shape == no_nbgs[["src_1", "src_2"]].values.flatten().shape

nbgs.to_pickle("data/all_nbgs.pkl")

no_nbgs.to_pickle("data/all_no_nbgs.pkl")
