==> legacy/06_sampling.ipynb <==
import sys
import os
import gc
import itertools as it
import tempfile
from collections import Counter, namedtuple
import multiprocessing as mp
import datetime as dt


import numpy as np
import pandas as pd

from joblib import Parallel, delayed, cpu_count, load, dump, Memory

%matplotlib inline
import matplotlib.pyplot as plt

# corral startup
sys.path.insert(1, "/home/jbcabral/carpyncho3")
os.environ.setdefault("CORRAL_SETTINGS_MODULE", "carpyncho.settings")

from corral import core
core.setup_environment()

from corral import db
from carpyncho.models import LightCurves
# from carpyncho.steps.features_extractor import CMP

from carpyncho.lib.beamc import (
    extinction, knnfix, add_columns, MIN_BOX_SIZE, SERVER_SOURCES_LIMIT)

with db.session_scope() as session:
    lc = session.query(LightCurves).filter(LightCurves.tile.has(name="b396")).one()
    feats = lc.features
    

feats = pd.DataFrame(feats)

nbgs = pd.read_pickle("data/all_nbgs.pkl")
no_nbgs = pd.read_pickle("data/all_no_nbgs.pkl")

src_ids = np.concatenate((
    np.unique(nbgs[["src_1", "src_2"]].values),
    np.unique(no_nbgs[["src_1", "src_2"]].values)))
src_ids = np.unique(src_ids)

feats = feats[feats.id.isin(src_ids)]

rows = []
for src_1, src_2 in nbgs[["src_1", "src_2"]].values:
    row = {"src1_{}".format(k): v for k, v in feats[feats.id == src_1].iloc[0].to_dict().items()}
    row.update({"src2_{}".format(k): v for k, v in feats[feats.id == src_2].iloc[0].to_dict().items()})
    row["nbg"] = 1
    rows.append(row)
    
for src_1, src_2 in no_nbgs[["src_1", "src_2"]].values:
    row = {"src1_{}".format(k): v for k, v in feats[feats.id == src_1].iloc[0].to_dict().items()}
    row.update({"src2_{}".format(k): v for k, v in feats[feats.id == src_2].iloc[0].to_dict().items()})
    row["nbg"] = 0
    rows.append(row)   
    
df = pd.DataFrame(rows)

columns = [c for c in list(df.columns) if c not in ['src2_ogle3_type', 'src1_ogle3_type', "nbg", 'src2_id', 'src1_id']]
columns = ["src1_id", "src2_id", "nbg"] + columns

df = df[columns]

# df.to_pickle("data/sample.pkl")

nbgs = pd.read_pickle("data/nbgs2.pkl")
no_nbgs = pd.read_pickle("data/no_nbgs2.pkl")

src_ids = np.concatenate((
    np.unique(nbgs[["src_1", "src_2"]].values),
    np.unique(no_nbgs[["src_1", "src_2"]].values)))
src_ids = np.unique(src_ids)

feats = feats[feats.id.isin(src_ids)]

rows = []
for src_1, src_2 in nbgs[["src_1", "src_2"]].values:
    row = {"src1_{}".format(k): v for k, v in feats[feats.id == src_1].iloc[0].to_dict().items()}
    row.update({"src2_{}".format(k): v for k, v in feats[feats.id == src_2].iloc[0].to_dict().items()})
    row["nbg"] = 1
    rows.append(row)
    
for src_1, src_2 in no_nbgs[["src_1", "src_2"]].values:
    row = {"src1_{}".format(k): v for k, v in feats[feats.id == src_1].iloc[0].to_dict().items()}
    row.update({"src2_{}".format(k): v for k, v in feats[feats.id == src_2].iloc[0].to_dict().items()})
    row["nbg"] = 0
    rows.append(row)   
    
df = pd.DataFrame(rows)

columns = [c for c in list(df.columns) if c not in ['src2_ogle3_type', 'src1_ogle3_type', "nbg", 'src2_id', 'src1_id']]
columns = ["src1_id", "src2_id", "nbg"] + columns

df = df[columns]

df.to_pickle("data/small_sample.pkl")
